Retrieving underlay criteria configs from back-end (#162)
* serving criteria configs with listUnderlays request + passing it appropriately to redux

* adding description for criteriaConfigs in open api underlay schema

* hard coding criteria configs into test underlay

* making hard coded strings private

* hard coding criteriaConfigs into fake underlays api

* making prettier formatting changes

* updating underlay

* adding criteria config property + builder

* setting criteria configs when building underlay

* removing hard coding of criteria config and setting it from underlay

* updating tests to support criteria configs

* putting criteria configs into underlay

* fixing criteria configs string in nautical underlay utils

* modifying fake underlay to have minimum configuration for passing test

* adding criteria configs to nautical test underlay

* wrapping criteria configs in quotes in underlays

* removing escaped quotes from criteria configs string

* removing unnecessary white space

* removing escape quotes from criteria configs string in test underlay

* adding comment for criteria configs property

* removing unnecessary imports

* formatting the criteria configs string

* adding indentation for criteria configs string
