+++ api/src/test/java/bio/terra/tanagra/service/underlay/NauticalUnderlayUtils.java
@@ -35,6 +35,8 @@ public final class NauticalUnderlayUtils {
   }
 
   public static final String NAUTICAL_UNDERLAY_NAME = "nautical_underlay";
+  public static final String NAUTICAL_CRITERIA_CONFIGS =
+      "[{\"type\":\"attribute\",\"title\":\"Age\",\"defaultName\":\"Age\",\"plugin\":{\"attribute\":\"age\"}]";
   public static final Entity SAILOR =
       Entity.builder().underlay(NAUTICAL_UNDERLAY_NAME).name("sailors").build();
   public static final Entity BOAT =
