+++ lib/src/main/proto/bio/terra/tanagra/underlay/underlay.proto
@@ -42,7 +42,9 @@ message Underlay {
 
   // The entity filters available on this underlay.
   repeated EntityFiltersSchema entity_filters_schemas = 8;
-
   // The entity indexing in the underlay.
   repeated EntityIndexing entity_indexing = 11;
+  
+  // The concepts and attributes in an underlay.
+  string criteria_configs = 12;
 }
