+++ ui/src/app.tsx
@@ -14,129 +14,6 @@ import "./app.css";
 
 enableMapSet();
 
-// TODO(tjennison): Fetch configs from the backend.
-const columns = [
-  { key: "concept_name", width: "100%", title: "Concept Name" },
-  { key: "concept_id", width: 120, title: "Concept ID" },
-  { key: "standard_concept", width: 180, title: "Source/Standard" },
-  { key: "vocabulary_id", width: 120, title: "Vocab" },
-  { key: "concept_code", width: 120, title: "Code" },
-];
-
-const criteriaConfigs = [
-  {
-    type: "concept",
-    title: "Conditions",
-    defaultName: "Contains Conditions Codes",
-    plugin: {
-      columns,
-      entities: [{ name: "condition", selectable: true, hierarchical: true }],
-    },
-  },
-  {
-    type: "concept",
-    title: "Procedures",
-    defaultName: "Contains Procedures Codes",
-    plugin: {
-      columns,
-      entities: [{ name: "procedure", selectable: true, hierarchical: true }],
-    },
-  },
-  {
-    type: "concept",
-    title: "Observations",
-    defaultName: "Contains Observations Codes",
-    plugin: {
-      columns,
-      entities: [{ name: "observation", selectable: true }],
-    },
-  },
-  {
-    type: "concept",
-    title: "Drugs",
-    defaultName: "Contains Drugs Codes",
-    plugin: {
-      columns,
-      entities: [
-        { name: "ingredient", selectable: true, hierarchical: true },
-        {
-          name: "brand",
-          sourceConcepts: true,
-          attributes: [
-            "concept_name",
-            "concept_id",
-            "standard_concept",
-            "concept_code",
-          ],
-          listChildren: {
-            entity: "ingredient",
-            idPath: "relationshipFilter.filter.binaryFilter.attributeValue",
-            filter: {
-              relationshipFilter: {
-                outerVariable: "ingredient",
-                newVariable: "brand",
-                newEntity: "brand",
-                filter: {
-                  binaryFilter: {
-                    attributeVariable: {
-                      variable: "brand",
-                      name: "concept_id",
-                    },
-                    operator: tanagra.BinaryFilterOperator.Equals,
-                    attributeValue: {
-                      int64Val: 0,
-                    },
-                  },
-                },
-              },
-            },
-          },
-        },
-      ],
-    },
-  },
-  {
-    type: "attribute",
-    title: "Ethnicity",
-    defaultName: "Contains Ethnicity Codes",
-    plugin: {
-      attribute: "ethnicity_concept_id",
-    },
-  },
-  {
-    type: "attribute",
-    title: "Gender Identity",
-    defaultName: "Contains Gender Identity Codes",
-    plugin: {
-      attribute: "gender_concept_id",
-    },
-  },
-  {
-    type: "attribute",
-    title: "Race",
-    defaultName: "Contains Race Codes",
-    plugin: {
-      attribute: "race_concept_id",
-    },
-  },
-  {
-    type: "attribute",
-    title: "Sex Assigned at Birth",
-    defaultName: "Contains Sex Assigned at Birth Codes",
-    plugin: {
-      attribute: "sex_at_birth_concept_id",
-    },
-  },
-  {
-    type: "attribute",
-    title: "Year at Birth",
-    defaultName: "Contains Year at Birth Values",
-    plugin: {
-      attribute: "year_of_birth",
-    },
-  },
-];
-
 // Prepackaged concept sets use _ in the ids to ensure they can't conflict with
 // user generated ones.
 const prepackagedConceptSets = [
@@ -203,11 +80,16 @@ export default function App() {
               throw new Error(`No entities in underlay ${name}`);
             }
 
+            const criteriaConfigs = res.underlays?.[i]?.criteriaConfigs;
+            if (!criteriaConfigs) {
+              throw new Error(`No criteria configs in underlay ${name}`);
+            }
+
             return {
               name,
               primaryEntity: "person",
               entities: entitiesRes.entities,
-              criteriaConfigs,
+              criteriaConfigs: JSON.parse(criteriaConfigs),
               prepackagedConceptSets,
             };
           })
